Migrate project to pnpm
