--- .yarnrc.yml
@@ -1,14 +1,0 @@
-approvedGitRepositories:
-  - '**'
-
-enableScripts: true
-
-nodeLinker: node-modules
-
-npmMinimalAgeGate: 0
-npmPreapprovedPackages:
-  - vitest
-  - '@vitest/*'
-catalog:
-  vite: npm:@voidzero-dev/vite-plus-core@0.2.8
-  vite-plus: 0.2.8
